Your commit message here

+++ src/app/compliance/compliance.component.ts
@@ -3,6 +3,7 @@ import { ComplianceModel } from './compliance.component.model';
 
 import { ApiService } from '../api.service';
 import { Router } from '@angular/router';
+import Swal from 'sweetalert2';
 
 
 @Component({
@@ -37,18 +38,37 @@ export class ComplianceComponent {
   }
 
   delete(id:number){
-    console.log('ID :::::',id );
-    this.apiService.deleteCompliance(id).subscribe(
-      (response:any)=>{
-        console.log(response.data);
-        window.location.reload();
-        alert("Record Deleted!");
-
-      },
-      (error:any)=>{
-        console.error(error);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!"
+      }).then((result) => {
+      if (result.isConfirmed) 
+      {
+        this.apiService.deleteCompliance(id).subscribe(
+          (response:any)=>{
+            console.log(response.data); 
+            Swal.fire({
+              title: "Record Deleted!",
+              icon: "success"
+            });
+          },
+          (error:any)=>{
+            console.error(error);
+            Swal.fire({
+              title: "Error!",
+              icon: "error"
+            });
+          }
+        );
+        setInterval(()=>{window.location.reload()},1000);        
       }
-    )
+    });
+    
   }
 
 applyFilter(): void {
